monero: models: refactor for cleaner project structure
