+++ src/main/java/org/hiahatf/mass/models/monero/proof/GetProofResult.java
@@ -1,4 +1,4 @@
-package org.hiahatf.mass.models.monero;
+package org.hiahatf.mass.models.monero.proof;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -6,13 +6,13 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 /**
- * POJO for the Monero transfer result.
+ * POJO for the get_reserve_proof result.
  * See Monero RPC docs for more details.
  */
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class GetReserveProofResult {
+public class GetProofResult {
     private String signature;
 }
