+++ src/main/java/org/hiahatf/mass/models/monero/proof/GetReserveProofRequest.java
@@ -1,4 +1,4 @@
-package org.hiahatf.mass.models.monero;
+package org.hiahatf.mass.models.monero.proof;
 
 import org.hiahatf.mass.models.Constants;
 
@@ -8,16 +8,16 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 /**
- * POJO for the Monero validate address request.
+ * POJO for the get_reserve_proof request.
  * See Monero RPC docs for more details.
  */
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class ValidateAddressRequest {
+public class GetReserveProofRequest {
     private final String jsonrpc = Constants.XMR_RPC_VER;
     private final String id = Constants.XMR_RPC_ID;
-    private final String method = Constants.XMR_RPC_VALIDATE_ADDRESS;
-    private ValidateAddressParameters params;
+    private final String method = Constants.XMR_RPC_GET_RESERVE_PROOF;
+    private GetReserveProofParameters params;
 }
