+++ src/main/java/org/hiahatf/mass/models/monero/proof/GetSpendProofParameters.java
@@ -1,4 +1,4 @@
-package org.hiahatf.mass.models.monero;
+package org.hiahatf.mass.models.monero.proof;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -6,15 +6,14 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 /**
- * POJO for the get_reserve_proof request.
+ * POJO for the get_spend_proof parameters.
  * See Monero RPC docs for more details.
  */
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class GetReserveProofParameters {
-    private boolean all;
-    private final int account_index = 0;
-    private long amount;
+public class GetSpendProofParameters {
+    private String txid;
+    private final String message = "MASS spend proof";
 }
