+++ src/main/java/org/hiahatf/mass/models/monero/proof/GetSpendProofRequest.java
@@ -1,4 +1,4 @@
-package org.hiahatf.mass.models.monero;
+package org.hiahatf.mass.models.monero.proof;
 
 import org.hiahatf.mass.models.Constants;
 
@@ -8,16 +8,16 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 /**
- * POJO for the Monero transfer request.
+ * POJO for the get_spend_proof request.
  * See Monero RPC docs for more details.
  */
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class TransferRequest {
+public class GetSpendProofRequest {
     private final String jsonrpc = Constants.XMR_RPC_VER;
     private final String id = Constants.XMR_RPC_ID;
-    private final String method = Constants.XMR_RPC_TRANSFER;
-    private TransferParameters params;   
+    private final String method = Constants.XMR_RPC_GET_SPEND_PROOF;
+    private GetSpendProofParameters params;
 }
