+++ src/main/java/org/hiahatf/mass/models/monero/transfer/TransferResponse.java
@@ -1,4 +1,4 @@
-package org.hiahatf.mass.models.monero;
+package org.hiahatf.mass.models.monero.transfer;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -6,15 +6,15 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 /**
- * POJO for the get_reserve_proof response.
+ * POJO for the transfer response.
  * See Monero RPC docs for more info.
  */
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class GetReserveProofResponse {
+public class TransferResponse {
     private String id;
     private String jsonrpc;
-    private GetReserveProofResult result;
+    private TransferResult result;
 }
