+++ src/main/java/org/hiahatf/mass/models/monero/validate/ValidateAddressRequest.java
@@ -1,0 +1,23 @@
+package org.hiahatf.mass.models.monero.validate;
+
+import org.hiahatf.mass.models.Constants;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+/**
+ * POJO for the validate_address request.
+ * See Monero RPC docs for more details.
+ */
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class ValidateAddressRequest {
+    private final String jsonrpc = Constants.XMR_RPC_VER;
+    private final String id = Constants.XMR_RPC_ID;
+    private final String method = Constants.XMR_RPC_VALIDATE_ADDRESS;
+    private ValidateAddressParameters params;
+}
