+++ src/main/java/org/hiahatf/mass/models/monero/validate/ValidateAddressResponse.java
@@ -1,0 +1,20 @@
+package org.hiahatf.mass.models.monero.validate;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+/**
+ * POJO for the validate_address response.
+ * See Monero RPC docs for more details.
+ */
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class ValidateAddressResponse {
+    private String id;
+    private String jsonrpc;
+    private ValidateAddressResult result;
+}
