+++ src/main/java/org/hiahatf/mass/models/monero/validate/ValidateAddressResult.java
@@ -1,4 +1,4 @@
-package org.hiahatf.mass.models.monero;
+package org.hiahatf.mass.models.monero.validate;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -6,7 +6,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 /**
- * POJO for the Monero validate address result.
+ * POJO for the validate_address result.
  * See Monero RPC docs for more details.
  */
 @Data
