+++ src/main/java/org/hiahatf/mass/services/monero/SwapService.java
@@ -7,10 +7,10 @@ import javax.net.ssl.SSLException;
 import org.hiahatf.mass.exception.MassException;
 import org.hiahatf.mass.models.Constants;
 import org.hiahatf.mass.models.lightning.InvoiceState;
-import org.hiahatf.mass.models.monero.TransferResponse;
 import org.hiahatf.mass.models.monero.SwapRequest;
 import org.hiahatf.mass.models.monero.SwapResponse;
 import org.hiahatf.mass.models.monero.XmrQuoteTable;
+import org.hiahatf.mass.models.monero.transfer.TransferResponse;
 import org.hiahatf.mass.repo.QuoteRepository;
 import org.hiahatf.mass.services.rpc.Lightning;
 import org.hiahatf.mass.services.rpc.Monero;
